requirements.txt
pyyaml
aiohttp
telegram
